// src/interfaces/interface.ts
export interface Product {
    id:number
    name?: string;
    price: number;
    imageUrl?: string;
    stock?: number;
    slogan?: string;
    desc?: string;
}
export interface Carts {
    id:number
    name?: string;
    price: number;
    imageUrl?: string;
    stock?: number;
    slogan?: string;
    desc?: string;
    quantity: number;
}